keep pagination component in one spot. some component refactoring
